refactor

+++ NeuToDo/ViewModels/SettingsViewModel.cs
@@ -15,7 +15,6 @@ namespace NeuToDo.ViewModels
         private readonly IPopupNavigationService _popupNavigationService;
         private readonly IAccountStorageService _accountStorageService;
         private readonly IDialogService _dialogService;
-        private readonly IBackupService _backupService;
         private readonly IEventModelStorage<NeuEvent> _neuStorage;
         private readonly IEventModelStorage<MoocEvent> _moocStorage;
         private readonly IDbStorageProvider _dbStorageProvider;
@@ -26,7 +25,6 @@ namespace NeuToDo.ViewModels
             IAccountStorageService accountStorageService,
             IDbStorageProvider dbStorageProvider,
             IDialogService dialogService,
-            IBackupService backupService,
             IContentPageNavigationService contentPageNavigationService,
             ICampusStorageService campusStorageService)
         {
@@ -36,10 +34,14 @@ namespace NeuToDo.ViewModels
             _accountStorageService = accountStorageService;
             _dbStorageProvider = dbStorageProvider;
             _dialogService = dialogService;
-            _backupService = backupService;
             _contentPageNavigationService = contentPageNavigationService;
             _campusStorageService = campusStorageService;
-            ServerPlatforms = ServerPlatform.ServerPlatforms;
+            accountStorageService.UpdateData += OnUpdateAccount;
+        }
+
+        private async void OnUpdateAccount(object sender, EventArgs e)
+        {
+            await LoadPlatforms();
         }
 
         private bool _isInit;
@@ -49,22 +51,34 @@ namespace NeuToDo.ViewModels
         private RelayCommand _pageAppearingCommand;
 
         public RelayCommand PageAppearingCommand => _pageAppearingCommand ??=
-            new RelayCommand(PageAppearingCommandFunction);
+            new RelayCommand(async () => await PageAppearingCommandFunction());
 
-        private void PageAppearingCommandFunction()
+        private async Task PageAppearingCommandFunction()
         {
             if (_isInit) return;
+
             Campus = _campusStorageService.GetCampus();
-            foreach (var platform in ServerPlatforms.Where(platform =>
-                _accountStorageService.AccountExist(platform.ServerType)))
+
+            await LoadPlatforms();
+
+            _isInit = true;
+        }
+
+        private async Task LoadPlatforms()
+        {
+            var defaultPlatforms = ServerPlatform.ServerPlatforms;
+
+            foreach (var platform in defaultPlatforms)
             {
-                platform.UserName = _accountStorageService.GetUserName(platform.ServerType);
-                platform.LastUpdateTime = _accountStorageService.GetUpdateTime(platform.ServerType);
+                var platformAccount = await _accountStorageService.GetAccountAsync(platform.ServerType);
+                if (platformAccount == null) continue;
+                platform.UserName = platformAccount.UserName;
+                platform.LastUpdateTime = platformAccount.LastUpdateTime;
                 platform.Button1Text = "更新";
                 platform.IsBound = true;
             }
 
-            _isInit = true;
+            ServerPlatforms = new List<ServerPlatform>(defaultPlatforms);
         }
 
         private RelayCommand<ServerPlatform> _command1;
@@ -82,42 +96,6 @@ namespace NeuToDo.ViewModels
         public RelayCommand<ServerPlatform> Command2 =>
             _command2 ??= new RelayCommand<ServerPlatform>(async (p) => await Command2Function(p));
 
-        private RelayCommand _importDb;
-
-        public RelayCommand ImportDb =>
-            _importDb ??= new RelayCommand(async () => await ImportDbFunction());
-
-        private async Task ImportDbFunction()
-        {
-            try
-            {
-                await _backupService.ImportAsync(new List<FileType> {FileType.Sqlite});
-                _dbStorageProvider.OnUpdateData();
-            }
-            catch (Exception e)
-            {
-                _dialogService.DisplayAlert("警告", e.ToString(), "Ok");
-            }
-        }
-
-        private RelayCommand _exportDb;
-
-        public RelayCommand ExportDb =>
-            _exportDb ??= new RelayCommand(async () => await ExportDbFunction());
-
-        private async Task ExportDbFunction()
-        {
-            try
-            {
-                var dest = await _backupService.ExportAsync();
-                _dialogService.DisplayAlert("提示", $"保存在{dest}", "OK");
-            }
-            catch (Exception e)
-            {
-                _dialogService.DisplayAlert("警告", e.ToString(), "Ok");
-            }
-        }
-
 
         private async Task Command2Function(ServerPlatform p)
         {
@@ -127,12 +105,14 @@ namespace NeuToDo.ViewModels
                 return;
             }
 
+            var res = await _dialogService.DisplayAlert("警告", "将清空数据库(未同步的数据将丢失)", "确认", "取消");
+            if (!res) return;
+
             var itemType = p.ServerType;
             _accountStorageService.RemoveAccountHistory(itemType);
             switch (itemType)
             {
                 case ServerType.Neu:
-
                     await _neuStorage.ClearTableAsync();
                     _dbStorageProvider.OnUpdateData();
                     break;
@@ -144,10 +124,7 @@ namespace NeuToDo.ViewModels
                     throw new ArgumentOutOfRangeException();
             }
 
-            p.UserName = string.Empty;
-            p.LastUpdateTime = string.Empty;
-            p.Button1Text = "关联";
-            p.IsBound = false;
+            _accountStorageService.OnUpdateData();
         }
 
         private RelayCommand _navigateToBackupPageCommand;
